+++ src/PKD.hs
@@ -6,7 +6,9 @@ License     : BSD-style
 Maintainer  : chrisreade@mac.com
 Stability   : experimental
 
-This is the main module to import Tgraphs and TileLib (which includes Halftile) modules.
+This is the main module to re-export both Tgraphs and TileLib (which includes Halftile) modules.
+However it does not export the data constructor Forced (only the newtype operator).
+It issues a warning when using makeUncheckedTgraph.
 -}
 -- {-# LANGUAGE NoMonomorphismRestriction #-}
 -- {-# LANGUAGE FlexibleContexts          #-}
@@ -17,29 +19,15 @@ This is the main module to import Tgraphs and TileLib (which includes Halftile) 
 module PKD ( module Tgraphs
            , module TileLib
            , makeUncheckedTgraph
-           , uncheckedCompose
-           , uncheckedPartCompose
            ) where
 
 
 import TileLib
-import Tgraphs hiding (makeUncheckedTgraph, uncheckedCompose, uncheckedPartCompose)
-import qualified Tgraphs as Unchecked (makeUncheckedTgraph, uncheckedCompose, uncheckedPartCompose)
+import Tgraphs hiding (makeUncheckedTgraph, Forced(Forced))
+import qualified Tgraphs as Unchecked (makeUncheckedTgraph)
 
 {-# WARNING makeUncheckedTgraph "Bypasses checks for required Tgraph properties. Use makeTgraph instead" #-}
 -- |Now issues a warning.
 -- Bypasses checks for required Tgraph properties. Use makeTgraph instead
 makeUncheckedTgraph :: [TileFace] -> Tgraph
 makeUncheckedTgraph = Unchecked.makeUncheckedTgraph
-
-{-# WARNING uncheckedCompose "Only safe when applied to a forced Tgraph. Use compose instead" #-}
--- |Now issues a warning
--- Only safe when applied to a forced Tgraph. Use compose instead
-uncheckedCompose :: Tgraph -> Tgraph
-uncheckedCompose = Unchecked.uncheckedCompose
-
-{-# WARNING uncheckedPartCompose "Only safe when applied to a forced Tgraph. Use partCompose instead" #-}
--- |Now issues a warning
--- Only safe when applied to a forced Tgraph. Use partCompose instead
-uncheckedPartCompose :: Tgraph -> ([TileFace], Tgraph)
-uncheckedPartCompose = Unchecked.uncheckedPartCompose